logger: Change the internal logger faults to more appropriate levels
The logger has a print for RecvError which was 0 but it's technically
"not an error", but something that can happen during normal operations,
thus, it gets a level of 1, which does not mean error outright, but
"this shouldn't happen besides shutdown".

Maybe a more appropriate level would be 32, halfway between error and
warning. Or maybe it should even be 128 (info). Hell why not, RecvError
happens only if the senders close their channels anyway.

<changed it to 128 instead of 1>

Now for dropped messages, that's easy, level 64 (warning). The reason
for this is that we do not expect the logger to ever get full during
normal operation, so it's useful to warn here.